Extract `Quintet` test environment into a separate class
